Add app environment and listen from multiple URLs
- Now the frontend can either be ran/built in development mode or in
production mode. This feature makes git pulling to the host machines
easier since they do not have to change URLs everytime

- The frontend can now listen to multiple URLs of the API.
We synchronously test each API URL since the API URLs all go to the same
machine and do not want to send multiple requests to the machine all
at once. The URL that is will be used the most should be at the front
of the URL list in the .env file

eg. URL_USED_THE_MOST;URL_USED_THE_LEAST

Note:
- may need to change the URLS to match the machine