Initial look for inputs of Advanced Search page
Due to time constraints,
here are the following limitation to the feature:
- We did not do the functionality for the export part of the
Advanced Search page since it was not too relevant to the inputs and
could be seperated into its own story
- Certain edge cases for the inputs are not handled
eg. File attachments for the products do not check whether the
file is actually a .txt or .csv file

The API calls that need to be updated for the new advanced search. Here
is a summary of the changes that needs to be done
(more details in the Advanced Search Doc)

- Include some parent-child relationship between categories and
subcategories without pulling the data from all the products
(either change GetAllSubCategoriesAsync or GetAllCategoriesAsync)

- Have an API call to get all the regions without needing to pull
the data from all the products

- Change Advanced Search API call to meet with the new inputs
(AdvanceSearchStoreProductsAsync)